requirements.txt
amqp==5.2.0
asgiref==3.7.2
Babel==2.14.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==42.0.0
Django==4.1.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-extra-views==0.14.0
django-haystack==3.2.1
django-localflavor==3.1
django-oscar==3.2.2
-e git+https://github.com/ayyaninam/fred-proj1-ANDAAL.git@b311672d14a8d723631b26972295fdf615e902ad#egg=django_oscar_paypal&subdirectory=django-oscar-paypal
-e git+https://github.com/ayyaninam/fred-proj1-ANDAAL.git@b311672d14a8d723631b26972295fdf615e902ad#egg=django_oscar_stripe_sca&subdirectory=django_oscar_stripe_sca
django-phonenumber-field==6.4.0
django-tables2==2.3.4
django-timezone-field==6.1.0
django-treebeard==4.7
django-widget-tweaks==1.5.0
factory-boy==3.2.1
Faker==22.2.0
idna==3.6
kombu==5.3.5
numpy==1.26.3
paypal-checkout-serversdk==1.0.3
paypalhttp==1.0.1
phonenumbers==8.13.27
pillow==10.2.0
prompt-toolkit==3.0.43
purl==1.6
pycparser==2.21
pycryptodome==3.20.0
pyOpenSSL==24.0.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-stdnum==1.19
pytz==2023.3.post1
rave-python==1.4.0
redis==5.0.1
requests==2.31.0
six==1.16.0
sorl-thumbnail==12.10.0
sqlparse==0.4.4
stripe==7.12.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
